Add files via upload
The legends for the land use and land cover map and the topographic map are presented in a unified manner. The vegetation map was created in accordance with the regulations delineated in the pertinent technical map manual. The legend was derived from the IBGE Environmental Information Bank, which can be accessed via the provided link. The Dynamic World map, in combination with its associated legend, was retrieved from the project's official platform.